producer/requirements.txt
kafka-python
websocket-client
python-dotenv
